AI_gen_image-master1/AI_gen_image-master/requirements.txt
customtkinter==5.2.0
pillow==10.0.0
requests==2.31.0
httpx==0.24.1
pytest==7.3.1
pyinstaller==6.0.0 